first working version of Resin launcher

+++ jeeunit-glassfish/src/main/java/com/googlecode/jeeunit/glassfish/EmbeddedGlassfishContainer.java
@@ -126,7 +126,9 @@ public class EmbeddedGlassfishContainer {
         Runtime.getRuntime().addShutdownHook(new Thread() {
             @Override
             public void run() {
-                tmpBeansXml.delete();
+                if (tmpBeansXml != null) {
+                    tmpBeansXml.delete();
+                }
                 shutdown();                
             }
         });
